class_t.restrict_sex is now a list of sex names to restrict (const char *).
`sex' is gone. everything is determined by gender which is simply
mlstring (and therefore, it depends on lang). Genders for
PC's do not have language-dependent parts (as short_descr's do).
OLC and mset should be working too.
No extensive testing was done though. The only things that were
checked is that old areas format is recognized, `asave world'
does not change anything unrelated to genders.

All the code that used char's sex for non-linguistic purposes now
checks that char is PC (i.e. his sex is not language-dependant).


git-svn-id: file:///home/fjoe/old-repos/svn/husky/trunk/sog@1877 1738b366-ca78-dd11-ac90-000e8e03bdb1