races and classes are fully copied in xxx_cpy. this should fix rename operations for them in OLC.
strkey_filename now generates filename with extension.

now condition checking rules does not have an order and are all
`deny' rules. each rule has a "main" predicate associated with it.
when rule is matched name of this predicate is returned by
`cc_vexpr_check' so the caller can distinguish the cause of denial.


git-svn-id: file:///home/fjoe/old-repos/svn/husky/trunk/sog@1959 1738b366-ca78-dd11-ac90-000e8e03bdb1