make[decl|util|dep].pl -> mk[decl|util|dep].pl.
_depend is not 'const char[]' (was const char *).

LOG_BUG added to use for logging incorrect control flow bugs.
LOG_ERROR should be used for logging errors which are not related to
software bugs (e.g.: file not found, incorrect keyword etc.)

`module_t.module_deps' implemented.
on reload of some module ("M") all the modules which depend on "M"
are added to set of modules to reload ("MODSET").
this set is checked for module validity and the modules are
unloaded/loaded in order of their `mod_prio'.
modules with higher `mod_prio' are unloaded last and loaded first.
note that module deps are read (and checked) only after the module
is loaded so modules that did not depend on "M"
but which become dependant on it at the moment when `mod_load' is
called (e.g.: their binaries were recompiled and updated)
do not get added to "MODSET".
initially (on startup) modules are loaded in order of their `mod_prio'
(because deps are not (completely) read yet at this stage) .

reverse varr iterators implemented (`varr_rforeach', `varr_reforeach',
`varr_rnforeach' and `varr_arnforeach').

use `MODULE' macro to load/unload cmds in _module_load/_module_unload
(MOD_XXX macros were used for some modules).


git-svn-id: file:///home/fjoe/old-repos/svn/husky/trunk/sog@2505 1738b366-ca78-dd11-ac90-000e8e03bdb1