Increase compatibility with our older area formats: - do not bamf out when unknown spec_xxx is found (log it with LOG_INFO instead of LOG_ERROR) - add dummy parsers for #MOBPROGS and #OMPROGS sections, skip old mobile mprog specification ('M')
Add an ability to load AL areas without editing them manually:
- skip AL "ManualRare" and "SkyVnum" keywords in area header
- skip AL 'R' (resists) mob extensions
- skip 'P' and 'V' (variables) obj extensions

AL areas are considered to be 'Ver 0' areas. No actual checks for sanity
of obj/mob flags and parameters which are saved in area in numeric (integer
or letter-flags) format were performed. They do not track area versions
so it is impossible to say whether they changed the meanings of such
numbers or not.


git-svn-id: file:///home/fjoe/old-repos/svn/husky/trunk/sog@4185 1738b366-ca78-dd11-ac90-000e8e03bdb1